+++ tests/test_graphstore.py
@@ -8,7 +8,7 @@ import json
 import httpretty
 from falcon import testing
 from graph_manager.app import init_api
-from graph_manager.utils.graph_store import GraphStore
+from graph_manager.applib.graph_store import GraphStore
 
 
 class GraphStoreTest(testing.TestCase):
